Starts hoisting JNI to plugin
